requirements.txt
requests
pillow
python-dotenv